Split components to separate files
